requirements.txt
llama-index==0.12.19
supabase
fastapi==0.115.8
openai==1.59.3
uvicorn==0.34.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.2
